\name{phtt-package}
\Rdversion{1.1}
\alias{phtt-package}
\alias{phtt}
\docType{package}
\title{
Panel Data Analysis with Heterogeneous Time Trends
}
\description{
  The Package \code{phtt} allows for the estimation of panel data models
  with an unobserved factor structure of the error term. This class of
  panel data models extends the classical within-models, which assume
  that the unobserved individual effects a constant over the time, and
  allows for the unobserved individual effects to vary with the time.
}
\details{
  \tabular{ll}{
    Package: \tab phtt\cr
    Type: \tab Package\cr
    Version: \tab 1.0\cr
    Date: \tab 2012-05-18\cr
    License: \tab GPL-2\cr
    LazyLoad: \tab yes \cr
  }
}
\author{
  Oualid Bada, 
  Dominik Liebl
}
\references{
  \itemize{
    \item Kneip, A., Sickles, R. C., Song, W., 2012 \dQuote{A New Panel Data
      Treatment for Heterogeneity in Time Trends}, \emph{Forthcoming in
      Econometric Theory}
    \item Bai, J., 2009 \dQuote{Panel data models with interactive fixed
      effects}, \emph{Econometrica}
}
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
\keyword{package, panel data, factor models}
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
% \examples{

% }
