\name{grid-internal}
\title{OptDim}
\alias{OptDim}
\alias{Eup}
\alias{checkSpecification}

\description{
  These Documentations still have to be written
}
\details{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written).
}
\keyword{ internal }
