\name{phtt-package}
\Rdversion{1.1}
\alias{phtt-package}
\alias{phtt}
\docType{package}
\title{
Panel Data Models for HTT
}
\description{
  The Package implements a new class of panel data models. This class
  allows for unobserved heterogeneity in time trends as suggested in .
  The package comes with diverse recently
  developed tests on dimensionality as well as with a routine
  for (functional) principal component analysis.  
}
\details{
  \tabular{ll}{
    Package: \tab phtt\cr
    Type: \tab Package\cr
    Version: \tab 1.0\cr
    Date: \tab 2010-03-31\cr
    License: \tab GPL-2\cr
    LazyLoad: \tab yes\cr
  }
}
\author{
  Oualid Bada
  Dominik Liebl
}
\references{
  \itemize{
    \item Bai, J., (2009) \dQuote{Panel Data Models with Interactive
      Fixed Effects}, \emph{Econometrica}, \bold{4}, 1229-1279
    \item Kneip, A., Sickles, R. C., Wonho, S., 2011 \dQuote{A New Panel Data
      Treatment for Heterogneity in Time Trends}, \emph{Forthcoming in
      Econometric Theory}
}
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
\keyword{package, panel data, factor models}
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
% \examples{

% }
