\name{grid-internal}
\alias{AFactMod}
\alias{Eup}
\alias{fpca.fit}
\alias{FUN.add.eff}
\alias{FUN.Pformula}
\alias{FUN.with.trans}
\alias{OptDim}
\title{OptDim}
\description{
  Internal phtt functions
}
\details{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written).
}
\keyword{ internal }
