\name{grid-internal}
\alias{AFactMod}
\alias{FUN.kss}
\alias{FUN.Pformula}
\alias{FUN.with.trans}
\alias{gafsa}
\title{Opt.dim}
\description{
  Internal phtt functions
}
\details{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written).
}
\keyword{ internal }
